+++ ctp2_code/gs/database/DiffDB.cpp
@@ -302,7 +302,7 @@ const char *DifficultyDB::GetYearStringFromTurn(sint32 diff, sint32 turn)
 	
 	static char buf[k_MAX_NAME_LEN];
 	if(g_useCustomYear && g_pTurnLengthOverride) {
-		if(turn >= g_turnLengthOverrideSize) {
+		if((unsigned) turn >= g_turnLengthOverrideSize) {
 			turn = g_turnLengthOverrideSize - 1;
 		}
 		strcpy(buf, g_pTurnLengthOverride[turn].text);
@@ -314,8 +314,8 @@ const char *DifficultyDB::GetYearStringFromTurn(sint32 diff, sint32 turn)
 		}
 
 		sprintf(buf, "%d %s", abs(year), ((year < 0) ?
-												 g_theStringDB->GetNameStr("str_tbl_ldl_BC") :
-												 g_theStringDB->GetNameStr("str_tbl_ldl_AD")));
+				g_theStringDB->GetNameStr("str_tbl_ldl_BC") :
+				g_theStringDB->GetNameStr("str_tbl_ldl_AD")));
 	}
 
 	return buf;
