+++ ctp2_code/ui/netshell/playereditwindow.cpp
@@ -601,7 +601,7 @@ void PlayerEditWindow::OKButtonAction::Execute(
 	aui_Radio *radio = (aui_Radio *)sg->GetChild( sg->WhichIsSelected() );
 	if ( radio )
 	{
-		for ( sint32 i = 0; i < sg->ChildList()->L(); i++ )
+		for ( sint32 i = 0; (unsigned) i < sg->ChildList()->L(); i++ )
 		if ( p->FindControl(
 			PlayerEditWindow::CONTROL_EXPERIENCE0CHECKBOX + i ) == radio )
 		{
